frametime 적용

+++ CrossyRoad 2.0/CrossyRoad 2.0/Struct.h
@@ -10,15 +10,15 @@ struct COLOR
 
 struct VECTOR2
 {
-	int x;
-	int y;
+	float x;
+	float y;
 };
 
 struct VECTOR3
 {
-	int x;
-	int y;
-	int z;
+	float x;
+	float y;
+	float z;
 };
 
 struct PLYINFO
